// lib/api/generated/model/chatbotUiSettingResponseDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Uchat API
 * Uchat API Documentation
 * OpenAPI spec version: 1.0
 */
import type { ChatbotAnimationConfigResponseDto } from './chatbotAnimationConfigResponseDto';
import type { ChatbotChatWindowConfigResponseDto } from './chatbotChatWindowConfigResponseDto';
import type { ChatbotFooterConfigResponseDto } from './chatbotFooterConfigResponseDto';
import type { ChatbotHeaderConfigResponseDto } from './chatbotHeaderConfigResponseDto';
import type { ChatbotInputConfigResponseDto } from './chatbotInputConfigResponseDto';
import type { ChatbotLauncherConfigResponseDto } from './chatbotLauncherConfigResponseDto';
import type { ChatbotMessageConfigResponseDto } from './chatbotMessageConfigResponseDto';
import type { ChatbotTypographyConfigResponseDto } from './chatbotTypographyConfigResponseDto';
import type { ChatbotUiSettingResponseDtoChatWindowPosition } from './chatbotUiSettingResponseDtoChatWindowPosition';
import type { ChatbotWelcomeConfigResponseDto } from './chatbotWelcomeConfigResponseDto';

export interface ChatbotUiSettingResponseDto {
  launcher: ChatbotLauncherConfigResponseDto;
  chat_window: ChatbotChatWindowConfigResponseDto;
  header: ChatbotHeaderConfigResponseDto;
  message: ChatbotMessageConfigResponseDto;
  typography: ChatbotTypographyConfigResponseDto;
  input: ChatbotInputConfigResponseDto;
  welcome: ChatbotWelcomeConfigResponseDto;
  animation: ChatbotAnimationConfigResponseDto;
  footer: ChatbotFooterConfigResponseDto;
  id: number;
  chatbot_id: number;
  primary_color: string;
  background_color: string;
  header_background_color: string;
  header_text_color: string;
  bot_message_background_color: string;
  bot_message_text_color: string;
  user_message_background_color: string;
  user_message_text_color: string;
  avatar_url?: string;
  logo_url?: string;
  launcher_icon_url?: string;
  header_title?: string;
  header_subtitle?: string;
  welcome_message: string;
  placeholder_text: string;
  chat_window_position: ChatbotUiSettingResponseDtoChatWindowPosition;
  chat_window_width: number;
  chat_window_height: number;
  border_radius: number;
  message_bubble_radius: number;
  show_avatar: boolean;
  show_logo: boolean;
}